Serve safari extension on our server
According to the instructions here:
https://developer.apple.com/library/safari/documentation/Tools/Conceptual/SafariExtensionGuide/DistributingYourExtension/DistributingYourExtension.html#//apple_ref/doc/uid/TP40009977-CH19-SW2

The .safariextz file needs to be served with the
`application/octet-stream` content type. Github serves this file with
the `text/html` content type. If we serve this from our server we have
control over the response type.

I'm hoping this will resolve the issue where the safari extension is
downloaded as a file instead of installed in safari when you click the
"Add to Safari" button.